Struct bgfx::Bgfx

pub struct Bgfx {
    // some fields omitted
}

BGFX object.

Contains all the calls intended to be run on the main thread. It is created through a call to bgfx::init(...), and will shut down BGFX when it is dropped.

Methods

impl Bgfx

fn reset(&self, width: u16, height: u16, reset: ResetFlags)

Resets the graphics device to the given size.

fn set_debug(&self, debug: DebugFlags)

Sets the debug flags in use.

fn set_view_clear(&self, id: u8, flags: ClearFlags, rgba: u32, depth: f32, stencil: u8)

Sets the options to use when clearing the given view.

fn set_view_rect(&self, id: u8, x: u16, y: u16, width: u16, height: u16)

Sets the rectangle to display the given view in.

fn set_view_transform(&self, id: u8, view: &[f32; 16], proj: &[f32; 16])

fn touch(&self, id: u8)

Touch a view. ( ͡° ͜ʖ ͡°)

fn dbg_text_clear(&self, attr: Option<u8>, small: Option<bool>)

Clears debug text.

fn dbg_text_image(&self, x: u16, y: u16, width: u16, height: u16, data: &[u8], pitch: u16)

Draws an image to the debug text overlay.

fn dbg_text_print(&self, x: u16, y: u16, attr: u8, text: &str)

Displays text in the debug text overlay.

fn frame(&self) -> u32

Finish the frame, syncing up with the render thread. Returns an incrementing frame counter.

fn set_transform(&self, mtx: &[f32; 16])

Sets the transform for rendering.

fn set_vertex_buffer(&self, vbh: &VertexBuffer)

Sets the vertex buffer for rendering.

fn set_index_buffer(&self, ibh: &IndexBuffer)

Sets the index buffer for rendering.

fn set_state(&self, state: StateFlags, rgba: Option<u32>)

Sets the render state.

fn submit(&self, view: u8, program: &Program) -> u32

Submit a primitive for rendering. Returns the number of draw calls.

fn get_renderer_type(&self) -> RendererType

Gets the type of the renderer in use.